remove_decimal

+++ database/migrations/2026_05_01_051341_create_detail_kurikulums_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->unsignedTinyInteger('semester');
             $table->unsignedTinyInteger('sesi_teori')->nullable();
             $table->unsignedTinyInteger('sesi_praktikum')->nullable();
-            $table->decimal('bobot_teori', 4, 2)->nullable();
-            $table->decimal('bobot_praktikum', 4, 2)->nullable();
+            $table->unsignedTinyInteger('bobot_teori')->nullable();
+            $table->unsignedTinyInteger('bobot_praktikum')->nullable();
             $table->enum('status_matkul', ['langsung', 'tidak langsung', 'pendukung']);
             $table->unsignedTinyInteger('sks');
